Converted the entire AiChatbot.jsx into an Modular Folder Structure for AiChatbot
