+++ Client/src/components/home/navBar/AiChatbot/styles/animations.css
@@ -1,0 +1,227 @@
+@keyframes float {
+  0%, 100% { transform: translateY(0px); }
+  50% { transform: translateY(-8px); }
+}
+
+@keyframes pulse-glow {
+  0%, 100% {
+    box-shadow: 0 4px 30px rgba(var(--shadow-rgb), 0.4),
+                0 0 60px rgba(var(--shadow-rgb), 0.2);
+  }
+  50% {
+    box-shadow: 0 4px 50px rgba(var(--shadow-rgb), 0.7),
+                0 0 80px rgba(var(--shadow-rgb), 0.4);
+  }
+}
+
+@keyframes gradient-shift {
+  0% { background-position: 0% 50%; }
+  50% { background-position: 100% 50%; }
+  100% { background-position: 0% 50%; }
+}
+
+@keyframes slideInUp {
+  from {
+    opacity: 0;
+    transform: translateY(20px);
+  }
+  to {
+    opacity: 1;
+    transform: translateY(0);
+  }
+}
+
+@keyframes morph-blob {
+  0%, 100% {
+    border-radius: 63% 37% 54% 46% / 55% 48% 52% 45%;
+    transform: scale(1) rotate(0deg);
+  }
+  25% {
+    border-radius: 40% 60% 54% 46% / 49% 60% 40% 51%;
+    transform: scale(1.05) rotate(90deg);
+  }
+  50% {
+    border-radius: 60% 40% 37% 63% / 61% 35% 65% 39%;
+    transform: scale(1.03) rotate(180deg);
+  }
+  75% {
+    border-radius: 33% 67% 53% 47% / 72% 31% 69% 28%;
+    transform: scale(1.07) rotate(270deg);
+  }
+}
+
+.shimmer-effect {
+  position: relative;
+  overflow: hidden;
+}
+
+.shimmer-effect::before {
+  content: '';
+  position: absolute;
+  inset: -2px;
+  background: linear-gradient(
+    45deg,
+    rgba(var(--shadow-rgb), 0.1),
+    rgba(var(--shadow-rgb), 0.4),
+    rgba(var(--shadow-rgb), 0.1),
+    rgba(var(--shadow-rgb), 0.3)
+  );
+  animation: morph-blob 8s ease-in-out infinite;
+  z-index: -1;
+}
+
+@keyframes particle-float {
+  0%, 100% { transform: translate(0, 0) rotate(0deg); opacity: 0.3; }
+  25% { transform: translate(10px, -10px) rotate(90deg); opacity: 0.6; }
+  50% { transform: translate(0, -20px) rotate(180deg); opacity: 0.3; }
+  75% { transform: translate(-10px, -10px) rotate(270deg); opacity: 0.6; }
+}
+
+@keyframes typing-dots {
+  0%, 20% { opacity: 0.2; }
+  50% { opacity: 1; }
+  100% { opacity: 0.2;
+  }
+}
+
+.ai-btn-wrapper {
+  position: relative;
+  display: inline-block;
+}
+
+.ai-btn-wrapper::before {
+  content: '';
+  position: absolute;
+  inset: -2px;
+  border-radius: 12px;
+  background: linear-gradient(45deg,
+    var(--btn),
+    var(--btn-hover),
+    var(--btn));
+  background-size: 200% 200%;
+  animation: gradient-shift 3s ease infinite;
+  opacity: 0.6;
+  filter: blur(8px);
+  z-index: -1;
+}
+
+.ai-ask-btn {
+  animation: float 3s ease-in-out infinite;
+  transition: all 0.3s cubic-bezier(0.68, -0.55, 0.265, 1.55);
+}
+
+.ai-ask-btn:hover {
+  transform: scale(1.05);
+}
+
+.chat-panel-wrapper {
+  animation: pulse-glow 3s ease-in-out infinite;
+}
+
+.message-bubble {
+  animation: slideInUp 0.3s ease-out;
+}
+
+.gradient-border {
+  position: relative;
+  border: 2px solid transparent;
+  background: var(--bg-ter);
+  background-clip: padding-box;
+}
+
+.gradient-border::before {
+  content: '';
+  position: absolute;
+  inset: 0;
+  border-radius: inherit;
+  padding: 2px;
+  background: linear-gradient(135deg,
+    var(--btn),
+    var(--btn-hover),
+    var(--btn));
+  -webkit-mask: linear-gradient(#fff 0 0) content-box,
+                linear-gradient(#fff 0 0);
+  -webkit-mask-composite: xor;
+  mask-composite: exclude;
+  animation: gradient-shift 3s linear infinite;
+  background-size: 200% 200%;
+}
+
+.particle {
+  position: absolute;
+  width: 4px;
+  height: 4px;
+  background: var(--btn);
+  border-radius: 50%;
+  pointer-events: none;
+  opacity: 0.3;
+}
+
+.particle:nth-child(1) { animation: particle-float 4s infinite; top: 10%; left: 10%; }
+.particle:nth-child(2) { animation: particle-float 5s infinite 0.5s; top: 20%; left: 80%; }
+.particle:nth-child(3) { animation: particle-float 6s infinite 1s; top: 70%; left: 20%; }
+.particle:nth-child(4) { animation: particle-float 4.5s infinite 1.5s; top: 80%; left: 70%; }
+.particle:nth-child(5) { animation: particle-float 5.5s infinite 2s; top: 40%; left: 50%; }
+
+.video-grid-bg {
+  position: absolute;
+  inset: 0;
+  background-image:
+    linear-gradient(rgba(var(--shadow-rgb), 0.03) 1px, transparent 1px),
+    linear-gradient(90deg, rgba(var(--shadow-rgb), 0.03) 1px, transparent 1px);
+  background-size: 20px 20px;
+  pointer-events: none;
+  opacity: 0.5;
+}
+
+.message-bold {
+  font-weight: 750;
+  color: inherit;
+}
+
+.message-italic {
+  font-style: italic;
+  color: inherit;
+}
+
+.glow-text {
+  text-shadow: 0 0 10px rgba(var(--shadow-rgb), 0.5),
+               0 0 20px rgba(var(--shadow-rgb), 0.3);
+}
+
+.input-glow:focus {
+  box-shadow: 0 0 0 3px rgba(var(--shadow-rgb), 0.2),
+              0 0 20px rgba(var(--shadow-rgb), 0.1);
+}
+
+.typing-dot {
+  animation: typing-dots 1.4s infinite;
+}
+
+.typing-dot:nth-child(2) {
+  animation-delay: 0.2s;
+}
+
+.typing-dot:nth-child(3) {
+  animation-delay: 0.4s;
+}
+
+.typing-cursor {
+  display: inline-block;
+  width: 2px;
+  height: 1em;
+  background-color: currentColor;
+  margin-left: 2px;
+  animation: blink 1s infinite;
+  vertical-align: middle;
+}
+
+@keyframes blink {
+  0%, 50% { opacity: 1; }
+  51%, 100% { opacity: 0; }
+}
+
+.message-content {
+  min-height: 1.5em;
+  line-height: 1.5;
+}
